Created pages folder with page components, created custom hook useSteppedForm and component SteppedFormControls
