+++ src/components/pages/CharacterSheet.js
@@ -1,17 +1,17 @@
 import React from "react";
-import { useAuth } from "../../../hooks/useAuth";
-import { useCharacterSheet } from "../../../hooks/characterSheet/useCharacterSheet";
-import { useFateRules } from "../../../hooks/useFateRules";
+import { useAuth } from "../../hooks/useAuth";
+import { useCharacterSheet } from "../../hooks/characterSheet/useCharacterSheet";
+import { useFateRules } from "../../hooks/useFateRules";
 
 import { Link } from "react-router-dom";
 import { Container, Button, Icon, Loader} from "semantic-ui-react"
 
-import CharacterMeta from "./CharacterMeta";
-import SheetPreview from "./SheetPreview";
-import { useCharacterBasics } from "../../../hooks/characterSheet/useCharacterBasics";
-import { useCharacterAspects } from "../../../hooks/characterSheet/useCharacterAspects";
-import { useCharacterStunts } from "../../../hooks/characterSheet/useCharacterStunts";
-import { useCharacterSkills } from "../../../hooks/characterSheet/useCharacterSkills";
+import CharacterMeta from "../character/sheet/CharacterMeta";
+import SheetPreview from "../character/sheet/SheetPreview";
+import { useCharacterBasics } from "../../hooks/characterSheet/useCharacterBasics";
+import { useCharacterAspects } from "../../hooks/characterSheet/useCharacterAspects";
+import { useCharacterStunts } from "../../hooks/characterSheet/useCharacterStunts";
+import { useCharacterSkills } from "../../hooks/characterSheet/useCharacterSkills";
 
 // This parent component (of all character sheet components) 
 // is used by the character sheet view (detail route)
